requirements.txt
python-whois==0.8.0
requests==2.31.0
tabulate==0.9.0 